Move mock cursor into shared location
